+++ spec/mobility/backends/active_record/table_spec.rb
@@ -265,6 +265,11 @@ describe "Mobility::Backends::ActiveRecord::Table", orm: :active_record, type: :
       backend_class.configure(options)
       expect(options[:foreign_key]).to eq(:article_id)
     end
+
+    it "sets composite foreign_key" do
+      backend_class.configure(options.merge!(foreign_key: [:article_id, :article_type]))
+      expect(options[:foreign_key]).to eq([:article_id, :article_type])
+    end
   end
 
   describe "with query plugin" do